test: 🚚 Move all inline tests into sibling tests.rs submodules
Pivots from the previous "tests/ at the top level" approach (which
forced one `pub fn` promotion of an internal helper) to the more
Rust-idiomatic submodule-via-`mod tests;` pattern. Every source file
now declares `#[cfg(test)] mod tests;` and the test code lives in a
sibling `tests.rs` file under `src/<module>/`.

Benefits over the tests/ approach:

- Tests retain access to private items via `use super::*` — no need to
  promote internal helpers to `pub` just to test them. Reverted
  `calculate_string_column_range` back to private.
- Tests stay logically co-located with the module they test (Rust
  idiom).
- `#[cfg(test)]` excludes test code from release builds.
- queries.rs and the other 10 library source files are now business-
  logic-only (no inline `mod tests {}` blocks diluting them).

Files restructured:

- src/queries.rs + src/queries/tests/{mod.rs, 7 test files}
- src/cache_manager.rs + src/cache_manager/tests.rs
- src/config.rs + src/config/tests.rs
- src/database.rs + src/database/tests.rs
- src/middleware_parser.rs + src/middleware_parser/tests.rs
- src/model_analyzer.rs + src/model_analyzer/tests.rs
- src/parser.rs + src/parser/tests.rs
- src/route_discovery.rs + src/route_discovery/tests.rs
- src/salsa_impl.rs + src/salsa_impl/tests.rs (flattens the original
  two `mod vite_tests` / `mod component_alias_tests` blocks)
- src/slot_navigation.rs + src/slot_navigation/tests.rs
- src/validation_rules.rs + src/validation_rules/tests.rs

The previously-created tests/ files (call_site_variants.rs and the
six pattern-extraction files) are moved into src/queries/tests/ as
proper submodules. tests/integration_tests.rs (the genuinely-
integration fixture-driven tests against a sample Laravel project)
stays where it belongs in tests/.

src/tests/ (the binary-target tests for LaravelLanguageServer methods
defined in main.rs) is unchanged — it's already a dedicated test
directory, and moving it would require migrating main.rs items into
lib.rs first.

Verification:
- cargo test: 129 lib + 64 main (src/tests/) + 79 integration_tests
  = 272 passing (unchanged from before the refactor)
- cargo check --tests: clean
- cargo clippy --lib: no new warnings introduced
- ~1400 lines of test code moved out of business-logic files into
  dedicated test submodules
